fix: prevent excessive 403 calls
When a project's details page is loaded, the CSI mechanism will request
the "Manage project" (aka edit) button.

However, since most folks are unlikely to be logged in, most of these
requests result in a 403 HTTP status code response.

Wrapping with an `if` statement, we should cut down the volume of
unnecessary requests to this endpoint significantly.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>