Audit/comment all uses of db.flush() and session.flush()
Calling flush() is _often_ unnecessary and can have unintended side effects
with some of our machinery. (Namely, I ran across confusing behavior when
beginning to add purging for organization objects.)

I've removed as many of the flushes as I was able in the codebase here, and did
my best to annotate the remaining ones.