test: wrap resend test case with a time freeze
One of the newer test cases using the `token_service` fixture wasn't
wrapped with the freeze decorator.
Other test cases in this module using the `token_service` **fixture**
are already wrapped as such, this follows suit.

There are some tests that use a stub token_service with guaranteed
`loads/dumps` behavior that don't use the fixture, so they don't need to
be frozen.

Resolves: #13233

Signed-off-by: Mike Fiedler <miketheman@gmail.com>