warehouse: abstract away publisher lookups
This devolves the lookup logic for each publisher type back
to its actual class (or base mixin, as appropriate).

The downside is that it's a little magical: each OIDCPublisher
now provides __lookup_strategies__, which is a list of callables
that resolve a model type and a claim set into an optional query,
which can then be realized with a DB session into exactly one
or none publishers.

Signed-off-by: William Woodruff <william@trailofbits.com>

tests: update tests

Signed-off-by: William Woodruff <william@trailofbits.com>