refactor: move the expcetion class somewhere safe
Due to the circular import nature of the OIDC modules, we cannot tell
pytest about the exception class without it traversing the import tree
to ensure it's there.

This isn't a very elegant solution, but it works, and drops the warning
noise emitted from this exception down to nothing, reducing ~1,300
warnings.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>