refactor: replace current windowed query
SQLAlchemy Wiki updated their suggested implementation circa 2020.

The existing code raises this deprecation warning:

```
RemovedIn20Warning: The Query.from_self() method is considered legacy
as of the 1.x series of SQLAlchemy and will be removed in 2.0.
The new approach is to use the orm.aliased() construct in conjunction
with a subquery.
See the section "Selecting from the query itself as a subquery" in the
2.0 migration notes for an example.
```

Instead of trying to update the syntax, use their suggested query which
emits the same behavior.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>