refactor: remove version from search documents
We refactored the way we select versions in #1519 when we provided a
`latest_version` key based on database query lookup at the time of
indexing.

The only usage was in the results.html template that no longer called
for `.version`, so we should be able to safely remove this key from the
search index, as well as the mechanism to populate it.

Found during efforts to refactor `windowed_query()` to SQLAlchemy 2.0
syntax.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>