test: remove pytest collect hook
We don't do anything with these markers, and the line between "unit" and
"functional" has gotten blurrier over time.

If we ever redefine and adhere to a standard, we can achieve the same
outcome by targeting a subdirectory for tests collection instead of
dynamically marking everything on collection phase.

Saves a couple of seconds per test invocation, as this logic doesn't
need to run for all of the test modules loaded.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>