chore: add guard to variable
`object_session()` can return None, but that shouldn't ever happen.
Without the guard, we get a type error when we call `sesssion.add()`
later on.

Instead of disabling the lint, add a guard. `no cover` is used since we
still don't have Event-model-only test cases to add that, and the risk
is low since our objects should always have sessions that retrieved
them.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>