chore(ci): fail tests if can't restore cache
With the refactor to make `deps` a common prerequisite, if we cannot
obtain the results from that step, there's no point in proceeding.

The failure of not having a cached environment results in a
semi-confusing error of "No module named ...", so failing when we cannot
restore should provide further clarity as to what happened.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>