refactor: include existing ultranormalized index
In #10498 and index was manually added via migration.
At the time, SQLAlchemy 1.4 did not detect this un-declared change.

Since SQLAlchemy 2.0, autogenerating a migration for the codebase wants
to drop the index, since it's not declared on the table.

Adding the correct column to the `__table_args__` means that the column
name needs to be in scope, which it cannot be at the top of the table
definition, so move it to the bottom, and add the existing Index.

With this change in place, alembic no longer suggests that we delete the
index, nor does it suggest any other operations.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>