feat(admin): add visibility for description
As an aid to debugging observed readme rendering issues, provide admins
a way to see what the current Description model columns contain.

Note: The "Raw Input"/"Stored Output" could get unwieldy, but they are
similar in nature to what's already being show in the "Metadata" section
of the same page.

Refs: #14064

Signed-off-by: Mike Fiedler <miketheman@gmail.com>