Step 2 of moving to Declarative syntax
In this step, we find&replace instances of `Column()` with
`mapped_column()` which presents an identical API.

The columns are still `Mapped[Any]` at this stage, those should be added
in Step 3.

Refs: https://docs.sqlalchemy.org/en/20/changelog/whatsnew_20.html#step-two-replace-declarative-use-of-schema-column-with-orm-mapped-column

Two Tables have not been converted to the declarative syntax, as they do
not have a primary key set yet. That needs to be handled out of band, as
there's data quality/migrations that need to happen first.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>