refactor: move verified email check into function
Introduced in #14126, check was only applied for the SessionSecurityPolicy.

Migrating it into the common helper applies to both Session and BasicAuth policies.

It was implicitly being checked before with BasicAuth via `_check_mfa`
which, if enabled and mandated, would have required a verified email
anyhow.

Moving it into the helper allows us to have a consistent requirement
across both policies.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>