chore: convert test_db models to declarative
The change in the expected output adopts `varchar` instead of `text`,
since Postgres doesn't care about which is which - a `varchar` with no
limit is the same as `text`. If for some reason we needed to be explicit,
we can always use the specific column type, but it's not needed
generally.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>