chore: enable tcp keepalive for celery/kombu/sqs
We perform a number of calls to SQS as part of normal operations -
sometimes on the order of 10 calls during an upload request.

Using TCP Keep-Alives is one way to speed up subsequent connections, and
the default is `False`.
Refs: https://boto3.amazonaws.com/v1/documentation/api/latest/guide/configuration.html#:~:text=respective%20regional%20endpoint.-,tcp_keepalive,-Toggles%20the%20TCP

Passing this config to Celery, which in turn sends it to Kombu, which
ion turn sends it to boto3, which sends to botocore, which sends to
urllib3, enables it.

That's quite a few turtles :turtle:

Signed-off-by: Mike Fiedler <miketheman@gmail.com>