refactor: update more test to string
Since Postgres treats `text` and `varchar` (with no limit) the same
under the hood, we don't need to make a distinction unless there's a
limit.

Note: There's still `Text` usage in `warehouse.packaging.models` which
has yet to be fully converted. Once it is, expect to see a similar
migration.

Refs: https://www.postgresql.org/docs/current/datatype-character.html

Signed-off-by: Mike Fiedler <miketheman@gmail.com>