chore: convert packaging models to PEP 484 Declarative
A behavior uncovered here that hasn't been dealt with before in this
codebase.
If the column in question is used in an existing trigger, we cannot
modify it. A possible change to support that would be to:

- drop the trigger
- perform the alter table
- recreate the trigger

I elected to not pursue this course right now, preferring to get as many
model fields in declarative format as possible.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>