refactor: make SearchLock a subclass
Instead of having our own implementation of lock acquisition, use the
one provided by `redis.lock.Lock`.

This didn't exist when we originally implemented SearchLock, but now it
does.

SearchLock sets a constant key name to be used for the lock,
and otherwise behaves the same.

Reduces the potential for race conditions where redis may have released
the lock, but the custom class had not yet due to timeout.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>