feat: enable sorted projects on user relationship
In order to obtain an ordered list of project names when calling
`User.projects`, refactor the existing relationship to use the more
modern `back_populates()` to express the relationship.

This allows us to implement the other side of the relationship as well
as express distinct properties for that side calling, specifically
`order_by()`.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>