fix: exist early on long search strings
Elasticsearch will raise an error if the string exceeds 1000 characters,
and we deliver a 5xx.

Instead of making ES do the work and try/catch, prevent the work from
being done at all if it's not going to succeed.

Most of these queries in recent times are paragraphs of text, which
isn't useful anyhow.

Refs: https://http.cat/413
Refs: https://github.com/elastic/elasticsearch/issues/91594
Refs: https://github.com/apache/lucene/blob/b292151f216424a65fdebbf0552e3c62c4505da7/lucene/core/src/java/org/apache/lucene/util/automaton/Operations.java#L1311

Signed-off-by: Mike Fiedler <miketheman@gmail.com>