chore: replace deprecated custom_predicates
`custom_predicates` was deprecated in Pyramid 1.5.
We don't actively use the ability we have to pass along more
`custom_predicates` and add them.

    DeprecationWarning: The "custom_predicates" argument to
    Configurator.add_route is deprecated as of Pyramid 1.5. Use
    "config.add_route_predicate" and use the registered route predicate as a
    predicate argument to add_route instead. See "Adding A Custom View,
    Route, or Subscriber Predicate" in the "Hooks" chapter of the
    documentation for more information.

Instead, construct a predicate factory that is then added to the route,
similar to `warehouse.predicates.DomainPredicate`.

I opted to keep this one close to its primary usage, instead of
relocating it to `warehouse.predicates`.

Refs: https://docs.pylonsproject.org/projects/pyramid/en/latest/narr/hooks.html#view-and-route-predicates

Signed-off-by: Mike Fiedler <miketheman@gmail.com>