chore: cache calls to signing key
With this simple caching mechanism, each running instance should only
have to make a single call at their first instantiation, and cache the
result for the lifetime of the process.

This call rarely fails, and adds ~200ms of each inbound hook, so
caching across requests should cut down the time it takes to complete
the processing.

Instead of using a Redis cache and worrying about cache expiration
strategies, if this ever fails a restart should evict the in-memory
cache and trigger a new HTTP call for the key.

Resolves #4463

Signed-off-by: Mike Fiedler <miketheman@gmail.com>