Warn users when API token is used in Trusted Publishing project
If a user uploads files using an API token to a project that has a
Trusted Publisher configured, send an email to all the project's users
notifying them of the API token use, suggesting they remove it and use
Trusted Publishing instead.

This warning is only sent once per (project, API token) combination.