test: remove db-controlled sequence from factory
When using `factory.Sequence()` for a database-controlled sequence
column, we can get into a weird state where we're reusing the sequence
since we've flushed already.

Since we already rely on the database to generate the ID for us, we
don't have to supply it ourselves as part of the factory. If we ever
*do* want to set a specific id for a JournalEntry, we would use a
keyword argument like `JournalEntryFactory.create(id=...)`

Exposed via:
--randomly-seed=2329742959

Signed-off-by: Mike Fiedler <miketheman@gmail.com>