feat: add captcha to login page
Mostly duplicates already-existing logic for `RegistrationForm`.
If we want to add the logic to a third place/form, then we should
refactor to a mixin at that time.

The hardest work here was updates to the test code, to enable a
"smarter" CSP service, since CSP is not implemented as an
interface/service.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>