feat: preserve related_name on Observations
If we want to allow ourselves to preserve some details about a given
Observation after the related model has been removed, copy some of the
data on to the Observation record.

Since we can infer the model type from the table name, we don't need to
store that.

Contains a data migration to update existing records and enforce the
requirement at the database level.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>