Fix GitLab Trusted Publishers UI and docs
Change the "workflow file path" field in the UI and in the docs to
"top-level pipeline file path". This is the correct term, since
GitLab OIDC claims only provide the top-level pipeline (usually
`.gitlab-ci.yml`), rather than individual jobs/workflows.

Also change the security model doc, to clarify that PyPI cannot
tell the difference between different .yml files that are included
in `.gitlab-ci.yml`, since it's the latter that is reported in
the claims.