Make TUF repository metadata update async
Make `tuf.update_metadata` an async task and call with `delay`.
The `project` argument becomes `project_id`, because it must be
serializable.

This change allows us to remove the `disable_tuf` test fixture, because
`delay` is mocked in the relevant tests anyway, so that
`update_metadata` isn't actually called in any existing tests. However,
call_recorder asserts need to be and are adapted.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>