Run notdatadog alongsite the tests
I found the test suite took ~30 minutes to run on my system, with CPU
and I/O curiously idle for the bulk of that period.

It turns out that they were furiously attempting to resolve the hostname
"notdatadog". As I'm working in a VM behind NAT they were getting
rate-limited by dnsmasq, which logged messages like this:

    Jul 16 23:16:24 vmhost dnsmasq[2230]: Maximum number of concurrent DNS queries reached (max: 150)

The quick fix is to run the notdatadog stub container alongside the
tests, which gives a runtime of ~1:30 on my system.