GitHub, GitLab: escape matchable characters during publisher lookup
This fixes a constraint violation that users were previously able
to induce by registering two similar-looking-but-not-exact
Trusted Publishers. For example:

```
repo: foo/bar
workflow: publish-pypi.yml
```

and:

```
repo: foo/bar
workflow: publish_pypi.yml
```

Before this change, PyPI's publisher lookup would treat
the `_` in `publish_pypi.yml` as a match expression,
resulting in both publishers being matched when only
exactly one publisher was expected to match. As a result,
a user could configure two publishers in such a way that
minting would *always* fail.

To fix this, we replace our use of a raw `like` with
a more constrained `startswith` *and* escape `_` and `%`
when they appear in the workflow identifier being matched on.
We would ideally do this using SQLAlchemy's `autoescape`,
but `autoescape` only supports string clauses and not
column clauses. Instead, we build up a somewhat verbose chain
of escape replacements manually.

Signed-off-by: William Woodruff <william@trailofbits.com>