lint: use type narrowing for `orm.object_session()`
Since `object_session(...)` can return `None`, we need to help mypy have
confidence that the session is indeed there.

One way to narrow the type is to assert that it's not `None`, and
therefore should not alert about `union-attr` problems.

Includes a couple of renames, and variable extraction where relevant.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>