Use `(test)pypi` in Trusted Publishing placeholder for GitHub Environments (#17036)
* Use `(test)pypi` in Trusted Publishing placeholder for GitHub Environments

GitHub Environments is a confusingly explained feature within GitHub that represents deployment targets. When projects get uploaded to PyPI — that a deployment target; same for TestPyPI. They don't represent processes but server-like entities. So using `release` is conceptually incorrect and gives people the wrong idea of what it is.

This is actually connected to Deployments API (and corresponding events) on the GitHub platform. The name Environments is just a misleading interface to describe Deployments that appears in some parts of the ecosystem, like GitHub Actions CI/CD.

In other places, it's called deployments and there's even a tab in repositories using it: https://github.com/cherrypy/cheroot/deployments/pypi. Each deployment can be linked to the corresponding released project version URL.

This patch attempts to align the practices with those used in the PyPUG guide and GitHub docs: https://github.com/actions/starter-workflows/pull/2345.

* Suggest `pypi` GitHub Environment @ `adding-a-publisher.md` doc

* Suggest `pypi` GitHub Environment @ `creating-a-project-through-oidc.md` doc

* Suggest `pypi` GitHub Environment @ `internals.md` doc

* Suggest `pypi` GitHub Environment @ `using-a-publisher.md` doc

* `make translations`

Signed-off-by: William Woodruff <william@trailofbits.com>

---------

Signed-off-by: William Woodruff <william@trailofbits.com>
Co-authored-by: William Woodruff <william@trailofbits.com>
Co-authored-by: Dustin Ingram <di@users.noreply.github.com>