Be slightly more strict in checking for distribution file contents
This improves our validation of distribution file contents by being more precise in checking for sdist's to have _one_ top level directory, this allows us to be more precise in checking for PKG-INFO and License-File files that should exist in a given sdist.

Also improves error messages for users if an invalid distribution is attempted to be uploaded.