chore(deps): replace vendored bourbon with package
The differences are minimal, and this sets an upgrade path to get us on
track for potential upgrades and removal.

After installing, using a creative `diff` command:

    diff --ignore-blank-lines --strip-trailing-cr \
        <(find warehouse/static/sass/tools/bourbon/ -type f -exec sed '/\/\*!/,/\*\//d' {} +) \
        <(find node_modules/bourbon/app/assets/stylesheets/ -type f -exec cat {} +)

The only differences found are:

- a header when vendored back in 5db84960b355b454ffb53e1b8376de8616e8feba
- license headers added in 9370bc5bf8ff847a2dbfe50f27773700c224d290
- spelling fixes in 13c5a7cac3ca0dae645f8f63076fab1686695bbb
- sass warning fixes in a0f3817879d952e6a5c1965d68bc933fac424974

Commits following this one will remove the vendored code, and start upgrades to
get us updated, and see about other removals.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>