Celery tuning (#17262)
* Celery tuning

Primarily focused on being able to explicitly set `--concurrency` in prod.

Unfortunately Celery doesn't respect an environment variable for this, so allow for setting additional Celery flags via the environment.

Also sets the dev worker to a concurrency of 1 to fix a locking issue with running Beat in the same process leading to many tasks being enqueued instead of one.

* specifically limit to --concurrency