perf: speed up user profile page
For users with a large amount of projects, the template iteration through
`projects` to fetch their latest `release` details turns into an n+1
query, and can time out on larger user profile pages.

Refactor the query to perform most of the logic to fetch only the data
necessary to render the template correctly, removing the need to load
full entities.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>