refactor: replace windowed_query with `yield_per`
Using a SQLAlchemy built-in for server-side cursor.

Instead of performing a single query at the top to create the windowed
columns, then issuing subsequent queries for windows of data, use
`yield_per()` to stream the results in a generator fashion using a
server-side cursor.

In dev, shaves ~40% of time off `reindex()`.

Refs: https://docs.sqlalchemy.org/en/20/orm/queryguide/api.html#fetching-large-result-sets-with-yield-per

Signed-off-by: Mike Fiedler <miketheman@gmail.com>