Improve error messages when a pending Trusted Publisher's project name already exists (#17405)
* Improve error messages when creating projects and publishers

This change improves the error messages when uploading a new project
or creating a pending Trusted Publisher, when the new project's name
already exists or is too similar to an existing project.

Signed-off-by: Facundo Tuesca <facundo.tuesca@trailofbits.com>

* Fix test error

Signed-off-by: Facundo Tuesca <facundo.tuesca@trailofbits.com>

* Fix name checking

Fix the check for projects with too-similar names when
there is more than one existing project with the same
ultranormalized name.

Signed-off-by: Facundo Tuesca <facundo.tuesca@trailofbits.com>

* Remove similar name from error messages

Signed-off-by: Facundo Tuesca <facundo.tuesca@trailofbits.com>

* Move errors back to interfaces.py

Signed-off-by: Facundo Tuesca <facundo.tuesca@trailofbits.com>

* De-parametrize tests

Signed-off-by: Facundo Tuesca <facundo.tuesca@trailofbits.com>

* Use exceptions instead of errors

Also change the TooSimilar error to store the conflicting
project name instead of a reference to the entire project.

Signed-off-by: Facundo Tuesca <facundo.tuesca@trailofbits.com>

* Address review comments

Signed-off-by: Facundo Tuesca <facundo.tuesca@trailofbits.com>

---------

Signed-off-by: Facundo Tuesca <facundo.tuesca@trailofbits.com>