feat: add index to Provenance.file_id
When used as a JOIN condition in `legacy.api.json.release` for the
release's files, the `provenance` table is consulted for any records
related to the `file_id`.

Currently unindexed, this is a sequential scan of the table of ~164k
rows, and ~87% of the total query cost.

Add an index to what is effectively a primary key to prevent further
slowdown as the provenance table grows with more projects publishing.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>