feat: compute top dependent corpus into cache (#17797)
* feat: compute top dependent corpus into cache

In preparation of being able to use a longer corpus, compute the top 10k
dependents and place them in the query results cache.

The results are not yet used, so this is a good way of exercising the
task logic without impeding any other processing.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* fix: account for periods in names and normalize

Also add a more complex specifier from the documentation.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* refactor: use inclusion vs exclusion

We don't care about `obsoletes`, `provides`, or legacy keys, only the
ones that are requirement dependencies on other projects.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* lint

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

---------

Signed-off-by: Mike Fiedler <miketheman@gmail.com>