Don't pre-compute the canonical URL
Turns out content within Jinja blocks will get executed even if they are
later replaced with another block. Additionally, if template rendering
fails for a exception view, it falls back to a plain status response.

Instead, use a request method to try and compute the canonical URL only
if one was not provided, and gracefully handle the case where the
canonical URL can't be constructed by not including a block at all.