Don't renegotiate possibly unknown default locale
Currently if the default locale negotiation fails, we fall back to
trying to match based on Accept-Language. If this fails, it falls back
on the default value, but because we re-negotiate the locale to
determine the default value (without determining if the default is a
known locale) we could potentially still set a garbage locale.

Since the default for the default_match in best_match is already `None`,
we can just use the default instead of re-negotiating.