feat: `request` shell helper
When working with the local interactive shell, it's often advantageous
to have a `request` like object to interact with the stack, similar to
the `db` and `config` objects.

Refactors test cases to replace `pretend` usages with `monkeypatch` and
`mocker` fixtures, and leverage more real fixtures like `pyramid_config`
while preserving the same level of assertions.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>