chore: exclude analytics tracking on some routes
Some behaviors like password resets include a token sent to the client's email.
When that token is used, the loaded webpage emits a tracking payload to
our private analytics service and is stripped prior to storage.

By excluding the tracking script from these pages, we prevent emitting
a tracking event on these pages that could contain the query string with the
token. We have other methods for tracking the traffic on these routes
without transmitting the token value.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>