feat(admin): quarantine all projects for user
When a user uploads many projects that all exhibit the same signature,
provide the admin user with the ability to quarantine and freeze all
projects for that user, as well as a single button to clear them from
quarantine.

Leverage existing project quarantine functionality, loop over
`user.projects` and pass each to be quarantined, which does the heavy
lifting.

Resolves #18403

Signed-off-by: Mike Fiedler <miketheman@gmail.com>