lint: enable unreachable code detection
Having mypy check for unreachable code during linting can prevent
accidental inclusion of code that cannot be reached, and that coverage
may not report.

This specific fix was surfaced by a check for `dm.user is None:` in
`macaroon/services.py` and since the type declaration for the `user`
column did not allow for None, the warning was raised as unreachable.

Since the database column `user_id` is allowed to have nulls, the
associated relationship may not always return a User, set it to match.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>