feat: pass the issuer_url to manage jkws storage (#18885)
* feat: pass the issuer_url to manage jkws storage

Previously using `self.issuer_url` didn't make a difference, as the
property is configured on the resolved Publisher class, and has nothing
to do with the inbound value.

Resolves #18845

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

* refactor: raise instead of passing none

Raise an exception if the key isn't found, handled by the caller chain
in `verify_jwt_signature()` prior to trying to decode a value with `None`.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>

---------

Signed-off-by: Mike Fiedler <miketheman@gmail.com>