fix: reverse index to journals to speed up triggers (#19056)
When a journal entry changes, the `update_project_last_serial` trigger
fires, which runs `maintain_project_last_serial()`.

This function looks for the `max(id)` for a given `name`, meaning it
first has to find all records for `name`, and then find the highest
`id` (at the "end" of the records).

By creating a compound index that has the name first, and then the most
recent `id` at the **top** of the index should return much faster,
speeding up any trigger/function executions, which are sadly somewhat
opaque to our instrumentation.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>