fix(sass): fix circular references
The migrator incorrectly selected some colours variables, this fixes
them, and resolves the erroring circular loading so now the modules are
loaded once, and do not have a circular dependency.

The remaining deprecation warnings are for fortawesome and its internal
behavior, which has changed to the modern approach in 7.x, which will be
upgraded separately.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>